IDM-1578: Refactor createSubAccount and getSubAccount
Thease methods will soon be used from another service.
